refactor(#266): split hotkey file to features
